+++ source/renderer/app/components/staking/delegation-center/DelegationCenterNoWallets.js
@@ -4,7 +4,6 @@ import { defineMessages, intlShape } from 'react-intl';
 import SVGInline from 'react-svg-inline';
 import { Button } from 'react-polymorph/lib/components/Button';
 import { ButtonSkin } from 'react-polymorph/lib/skins/simple/ButtonSkin';
-
 import styles from './DelegationCenterNoWallets.scss';
 import icon from '../../../assets/images/attention-big-thin.inline.svg';